% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datageneration.R
\name{datageneration}
\alias{datageneration}
\title{Linear normal data generation}
\usage{
datageneration(n, var, mean, sd)
}
\arguments{
\item{n}{A numeric element or vector denoting the number of observations.}

\item{var}{A numeric elemnt of vector denoting the number of variables.}

\item{mean}{A list denoting the means of the variables.}

\item{sd}{A list denoting the standard deviations of the variables.}
}
\value{
Returns a data set containing the independent variables and the generated dependent variable, y, which is linearly generated
}
\description{
Linear normal data generation
}
\examples{
datageneration(100, 3, list(1, 2, 3), list(1, 1, 1))
}
